20220407 agora commit

+++ lib/providers/group_notice.dart
@@ -9,12 +9,13 @@ class GroupNoticeProvider with ChangeNotifier {
   UserNoticeService _userNoticeService = UserNoticeService();
 
   Future<bool> create({
-    required String groupId,
-    required String title,
-    required String message,
+    String? groupId,
+    String? title,
+    String? message,
   }) async {
-    if (groupId == '') return false;
-    if (title == '') return false;
+    if (groupId == null) return false;
+    if (title == null) return false;
+    if (message == null) return false;
     try {
       String _id = _groupNoticeService.id(groupId);
       _groupNoticeService.create({
@@ -32,11 +33,15 @@ class GroupNoticeProvider with ChangeNotifier {
   }
 
   Future<bool> update({
-    required String id,
-    required String groupId,
-    required String title,
-    required String message,
+    String? id,
+    String? groupId,
+    String? title,
+    String? message,
   }) async {
+    if (id == null) return false;
+    if (groupId == null) return false;
+    if (title == null) return false;
+    if (message == null) return false;
     try {
       _groupNoticeService.update({
         'id': id,
@@ -51,35 +56,46 @@ class GroupNoticeProvider with ChangeNotifier {
     }
   }
 
-  void delete({required GroupNoticeModel groupNotice}) {
-    _groupNoticeService.delete({
-      'id': groupNotice.id,
-      'groupId': groupNotice.groupId,
-    });
+  Future<bool> delete({
+    String? id,
+    String? groupId,
+  }) async {
+    if (id == null) return false;
+    if (groupId == null) return false;
+    try {
+      _groupNoticeService.delete({
+        'id': id,
+        'groupId': groupId,
+      });
+      return true;
+    } catch (e) {
+      print(e.toString());
+      return false;
+    }
   }
 
   Future<bool> send({
-    required String id,
-    required String groupId,
-    required String title,
-    required String message,
-    required List<UserModel> users,
+    GroupNoticeModel? notice,
+    List<UserModel>? users,
   }) async {
+    if (notice == null) return false;
+    if (users == null) return false;
+    if (users.length == 0) return false;
     try {
       for (UserModel _user in users) {
         _userNoticeService.create({
-          'id': id,
-          'groupId': groupId,
+          'id': notice.id,
+          'groupId': notice.groupId,
           'userId': _user.id,
-          'title': title,
-          'message': message,
+          'title': notice.title,
+          'message': notice.message,
           'read': false,
           'createdAt': DateTime.now(),
         });
         _userNoticeService.send(
           token: _user.token,
-          title: title,
-          body: message,
+          title: notice.title,
+          body: notice.message,
         );
       }
       return true;
